reuse rsa's le_bits_to_bytes_targets method for ed25519pod too
le_bits_to_bytes_targets was implemented by @ndwrobotics at PR #7, and
it reduces the amount of gates used compared to original
bits_to_bytes_targets method. Now both Ed25519Pod and RsaPod use the
method.

The observed reduction is 2x improvement:
for 864 bits, goes from 76 to 38 gates
for 256 bits, goes from 22 to 11 gates
